Внесены правки: добавлены модули data и str_helper, изменена генерация Last Name
