+++ data/data.py
@@ -1,0 +1,7 @@
+"""
+Модуль содержит статические тестовые данные.
+"""
+GENERAL_URL = 'https://www.globalsqa.com/angularJs-protractor/BankingProject/#/manager'
+ADD_CUST_URL = 'https://www.globalsqa.com/angularJs-protractor/BankingProject/#/manager/addCust'
+CUSTOMERS_URL = 'https://www.globalsqa.com/angularJs-protractor/BankingProject/#/manager/list'
+ALERT_MESSAGE = 'Customer added successfully with customer id'
